// eshop-main/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type CartItem = {
  __typename?: 'CartItem';
  id: string;
  productId: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  product?: Maybe<Product>;
};

export type Cart = {
  __typename?: 'Cart';
  id: string;
  subtotal: number;
  total: number;
  items: CartItem[];
};
// ...
export type Scalars = {
  ID: string;
  String: string;
  Int: number;
  Float: number;
  Boolean: boolean;
  JSON: Record<string, unknown>;
};
// ...
export type GetCartQuery = { cart?: Maybe<Cart> };

export type GetCartQueryVariables = Exact<{ [key: string]: never }>;
// ...
export type RemoveFromCartMutation = { removeFromCart: Cart };

export type RemoveFromCartMutationVariables = Exact<{ itemId: Scalars['String'] }>;
// ...
export const GetCartDocument = gql`
  query GetCart {
    cart {
      id
      subtotal
      total
      items {
        id
        productId
        quantity
        unitPrice
        totalPrice
        product {
          id
          name
          thumbnailUrl
        }
      }
    }
  }
` as unknown as TypedDocumentNode<GetCartQuery, GetCartQueryVariables>;
// ...
export const RemoveFromCartDocument = gql`
  mutation RemoveFromCart($itemId: String!) {
    removeFromCart(input: { itemId: $itemId }) {
      id
      subtotal
      total
      items {
        id
        productId
        quantity
        unitPrice
        totalPrice
      }
    }
  }
` as unknown as TypedDocumentNode<RemoveFromCartMutation, RemoveFromCartMutationVariables>;